+++ include/chromadec/chromadec.h
@@ -11,6 +11,7 @@
 #include <chromadec/frame.h>
 #include <chromadec/nn.h>
 #include <chromadec/dropout.h>
+#include <chromadec/calibration.h>
 #include <chromadec/pipeline.h>
 
 #endif